pyproject.toml
[tool.poetry]
name = "video-watermark-decoder"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
opencv-python = "^4.5.5"
matplotlib = "^3.5.1"
notebook = "^6.4.11"
pydub = "^0.25.1"
pandas = "^1.4.2"
scipy = "^1.8.0"
numpy = "^1.22.3"
bitarray = "^2.5.1"
bchlib = "^0.14.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
